backend/requirements.txt
async-timeout==4.0.3
bidict==0.23.1
blinker==1.7.0
Cerberus==1.3.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
crc==6.1.1
dnspython==2.6.1
exceptiongroup==1.2.0
Faker==24.3.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.4.0
pymongo==4.6.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-redis==0.4.0
python-socketio==5.11.1
redis==5.0.3
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
tomli==2.0.1
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
